Update LifeSpanHandler.CS
I added a condition, which will ensure that if the popup is a 'Sign In or Sign Up with Google' window, it will use CEFSharp Default popup. I noticed that the myForm.AddNewBrowserTab method will not help users to sign in with google. Therefore I used if and else statements to use Default CefSharp popup only if the popup's url has mention of google.